Fix utils
